+++ test/test.ml
@@ -52,10 +52,10 @@ let test_exn_typechecks () =
   let (_ : bool) = true || Fmt.invalid_arg "%s" "" in
   ()
 
-let test_kstrf_strf_like_partial_app () =
+let test_kstr_str_like_partial_app () =
   let assertf f = assert (f "X" = f "X") in
-  let test_kstrf fmt = Fmt.kstrf (fun x -> x) fmt in
-  let test_strf_like fmt = Fmt.strf_like Fmt.stderr fmt in
+  let test_kstrf fmt = Fmt.kstr (fun x -> x) fmt in
+  let test_strf_like fmt = Fmt.str_like Fmt.stderr fmt in
   assertf (test_strf_like "%s");
   assertf (test_kstrf "%s");
   ()
@@ -65,7 +65,7 @@ let tests () =
   test_dump_uchar ();
   test_utf_8 ();
   test_style_renderer ();
-  test_kstrf_strf_like_partial_app ();
+  test_kstr_str_like_partial_app ();
   Printf.printf "Done.\n";
   ()